agents/ocr_agent/requirements.txt
# python web framework
fastapi==0.111.0

# Uvicorn is the server that runs FastAPI
uvicorn==0.29.0

# Pytesseract is the Python wrapper around Tesseract OCR
pytesseract==0.3.10

# Pillow is a Python image processing library
Pillow==10.3.0

# converts PDF pages into images so Tesseract can read them
pdf2image==1.17.0

# python-multipart allows FastAPI to receive uploaded files
python-multipart==0.0.9